moviesHaven/viewsets.py
from moviesHaven.models import *
from moviesHaven.serializers import *
from rest_framework import viewsets


class MovieViewSet(viewsets.ModelViewSet):
    queryset = Movie.objects.all()
    serializer_class = MovieSerializer
