requirements.txt
Django==2.0.2
djangorestframework==3.7.7
django-filter==1.1.0
Markdown==2.6.11
mysqlclient==1.3.12
Pillow==5.0.0
requests==2.18.4
